requirements.txt

flask==1.1.4
Werkzeug==1.0.1
markupsafe==2.0.1
Jinja2==2.10.1
click==7.1.2
itsdangerous==1.1.0
# charset_normalizer==3.1.0
fake-useragent==1.1.3
pandas==1.1.4
flask_sqlalchemy==2.5.1
flask_babelex==0.9.4
flask_admin==1.6.1
sqlalchemy==1.4.32
Sphinx==1.8.0
selenium==3.141.0
bs4==0.0.1
# wtforms==3.0.1
jieba
flask_script

numpy
snownlp
pyecharts
wordcloud
pymysql
pillow
requests
gensim
nltk




